+++ extension/popup/interfaces.d.ts
@@ -4,3 +4,14 @@ export interface SearchItem extends HTMLElement {
         lng: string;
     }
 }
+export interface SavedPreferences {
+    latitude: number;
+    longitude: number;
+    enabled: boolean;
+    /** Map latest zoom level */
+    zoom?: number;
+    /** Nominatim location name */
+    locationName?: string;
+    /** User selected distance for randomization */
+    distanceIndex: number;
+}